app/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.6.2.post1; python_version >= '3.9'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
boto3==1.35.40; python_version >= '3.8'
botocore==1.35.40; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
dnspython==2.7.0; python_version >= '3.9'
email-validator==2.2.0
exceptiongroup==1.2.2; python_version < '3.11'
fastapi[standard]==0.115.2; python_version >= '3.8'
fastapi-cli[standard]==0.0.5
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.6; python_version >= '3.8'
httptools==0.6.2
httpx==0.27.2
idna==3.10; python_version >= '3.6'
jinja2==3.1.4
jmespath==1.0.1; python_version >= '3.7'
lxml==5.3.0; python_version >= '3.6'
mangum==0.19.0; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.1; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.12
pyyaml==6.0.2
rich==13.9.2; python_full_version >= '3.8.0'
s3transfer==0.10.3; python_version >= '3.8'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
starlette==0.40.0; python_version >= '3.8'
typer==0.12.5; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.10'
uvicorn[standard]==0.31.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
